Refactor server management and world information services
- Remove MinecraftInfoService and consolidate world retrieval logic in WorldManagementService
- Update MinecraftServerController to use simplified world and properties retrieval methods
- Add caching for player and server properties information
- Enhance ServerProcessManager with more robust server discovery and management
- Simplify MinecraftWorld DTO by removing redundant properties
- Update PlayerDto to include additional player attributes
