+++ src/main/java/com/mcap/minecraftagent/service/ServerPropertiesService.java
@@ -1,7 +1,10 @@
 package com.mcap.minecraftagent.service;
 
+import com.mcap.minecraftagent.dto.ServerPropertiesDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
@@ -41,7 +44,13 @@ public class ServerPropertiesService {
         return properties;
     }
 
-    public Map<String, String> getAllProperties(String worldId) {
+    @Cacheable(value = "propertiesInfo", key = "#worldName")
+    public ServerPropertiesDto getProperties(String worldName) {
+        log.info("Fetching properties for world: {}", worldName);
+        return fetchProperties(worldName);
+    }
+
+    public ServerPropertiesDto fetchProperties(String worldId) {
         log.info("Retrieving all properties for worldId: {}", worldId);
         Properties properties = loadProperties(worldId);
         Map<String, String> propsMap = new HashMap<>();
@@ -49,7 +58,7 @@ public class ServerPropertiesService {
             propsMap.put(key, properties.getProperty(key));
         }
         log.info("Successfully retrieved {} properties for worldId: {}", propsMap.size(), worldId);
-        return propsMap;
+        return new ServerPropertiesDto(propsMap);
     }
 
     public void setProperty(String worldId, Map<String, String> properties) {
@@ -68,10 +77,7 @@ public class ServerPropertiesService {
         }
         log.info("Finished updating properties for worldId: {}", worldId);
         saveProperties(propertiesPath, propFile);
-        var cache = this.cacheManager.getCache("minecraftWorlds");
-        if (cache != null) {
-            cache.clear();
-        }
+        evictPropertiesCache(worldId);
     }
 
     private void saveProperties(String propertiesPath, Properties properties) {
@@ -84,5 +90,10 @@ public class ServerPropertiesService {
             log.error("Failed to save properties file: {}", propertiesPath, e);
             throw new RuntimeException("Failed to save properties file: " + propertiesPath, e);
         }
+    }
+
+    @CacheEvict(value = "propertiesInfo", key = "#worldName")
+    public void evictPropertiesCache(String worldName) {
+        log.info("Evicting properties cache for world: {}", worldName);
     }
 }
